[fix]- método update

+++ AgendaUoW/Domain/Services/IContatoService.cs
@@ -7,7 +7,7 @@ namespace AgendaUoW.Domain.Services
     public interface IContatoService
     {
         Task<Contato> Salvar(Contato contato);
-        Task<Contato> Editar(int idContato, Contato contato);
+        Task<Contato> Editar(decimal idContato, Contato contato);
         Task<bool> Excluir(decimal idContato);
         Task<IEnumerable<Contato>> ObterPorNome(string nome);
         Task<IEnumerable<Contato>> ObterPorNumero(string numero);
